+++ installer/ubuntu/scripts/resize.sh
@@ -23,6 +23,6 @@ export LOOP_BOOT="${LOOP}p1"
 echo "+ loop is $LOOP"
 
 RESIZE_CMD="e2fsck -f $LOOP_ROOT; e2fsck -f $LOOP_BOOT; resize2fs $LOOP_ROOT"
-DOCKER_DEFAULT_PLATFORM="" docker pull ubuntu:23.10
-DOCKER_DEFAULT_PLATFORM="" docker run --privileged -v $LOOP_ROOT:$LOOP_ROOT ubuntu:23.10 sh -c "$RESIZE_CMD"
+DOCKER_DEFAULT_PLATFORM="" docker pull ubuntu:24.04
+DOCKER_DEFAULT_PLATFORM="" docker run --privileged -v $LOOP_ROOT:$LOOP_ROOT ubuntu:24.04 sh -c "$RESIZE_CMD"
 losetup -d $LOOP 2>/dev/null